requirements.txt
streamlit
openai==1.12.0
langchain==0.1.17
langchain-openai==0.0.8
python-dotenv